# contracts/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API for the Reverse Marketplace platform",
    "title": "Reverse Marketplace API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/admin/disputes/{disputeId}/resolve": {
      "post": {
        "description": "Resolve dispute",
        "parameters": [
          {
            "description": "Dispute UUID",
            "in": "path",
            "name": "disputeId",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "outcome": {
                    "description": "Resolution outcome",
                    "enum": [
                      "full_refund",
                      "partial_refund",
                      "no_refund",
                      "custom"
                    ],
                    "type": "string"
                  },
                  "refundAmount": {
                    "description": "Refund amount for partial_refund/custom (e.g. 50.00)",
                    "minimum": 0,
                    "type": "number"
                  },
                  "resolutionSummary": {
                    "description": "Summary of the resolution decision (min 10 chars)",
                    "maxLength": 2000,
                    "minLength": 10,
                    "type": "string"
                  }
                },
                "required": [
                  "outcome",
                  "resolutionSummary"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Reverse Marketplace API (base path)",
      "url": "/api/v1"
    }
  ]
}